Move BruteForce struct to BruteForceOperation class.
